requirements.txt
numpy
scipy
matplotlib
scikit-learn
joblib
GitPython

tensorflow==1.8.0
pydart2
pybullet
